packages/stream/pyproject.toml
[project]
name = "behavior-stream"
version = "1.0.0"
description = "Stream processing service for BehaviorSense"
requires-python = ">=3.11"

dependencies = [
    "behavior-core",
    "faust-streaming>=0.11.0",
    "pulsar-client>=3.4.0",
    "clickhouse-connect>=0.6.0",
    "redis[asyncio]>=5.0.0",
]

[tool.uv.sources]
behavior-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/behavior_stream"]
